// webapp/src/app/interface/accessibility.interface.ts
export interface Accessibility {
  id: string;
  toiletGrabBarAvailable: boolean;
  showerGrabBarAvailable: boolean;
  stepFreeShowerAvailable: boolean;
  showerBathChairAvailable: boolean;
  stepFreeBedroomAccessAvailable: boolean;
  wideBedroomEntranceAvailable: boolean;
  stepFreeAccessAvailable: boolean;
}